fix: 컬렉션 관련 로직 더 잘 보이게 컬렉션 컨테이너로 이동
